src/config.py
# API 키, 환경변수 등 설정값을 관리하는 파일
import os
from dotenv import load_dotenv

# .env 파일에서 환경변수 읽기
load_dotenv()

# 한국투자증권 API 설정
KIS_APP_KEY = os.getenv("KIS_APP_KEY", "")
KIS_APP_SECRET = os.getenv("KIS_APP_SECRET", "")
KIS_ACCOUNT_NO = os.getenv("KIS_ACCOUNT_NO", "")

# 한국투자증권 API 엔드포인트
KIS_DOMAIN = "https://openapi.koreainvestment.com:9443"  # 실전 환경
# KIS_DOMAIN = "https://openapivts.koreainvestment.com:29443"  # 모의 환경
